+++ core/src/main/java/es/urjc/etsii/grafo/io/serializers/excel/ExcelSerializer.java
@@ -1,6 +1,7 @@
 package es.urjc.etsii.grafo.io.serializers.excel;
 
 import es.urjc.etsii.grafo.io.Instance;
+import es.urjc.etsii.grafo.io.InstanceManager;
 import es.urjc.etsii.grafo.io.serializers.ResultsSerializer;
 import es.urjc.etsii.grafo.solution.Solution;
 import es.urjc.etsii.grafo.solver.SolverConfig;
@@ -15,9 +16,12 @@ import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.logging.Logger;
+
+import static es.urjc.etsii.grafo.io.serializers.excel.RawSheetWriter.writeCell;
 
 
 /**
@@ -38,6 +42,11 @@ public class ExcelSerializer<S extends Solution<S,I>, I extends Instance>  exten
     public static final String PIVOT_SHEET = "Pivot Table";
 
     /**
+     * Instance data sheet
+     */
+    public static final String INSTANCE_SHEET = "Instances";
+
+    /**
      * Positive infinity value to use in Excel
      */
     public static final double POSITIVE_INFINITY = 1e99;
@@ -51,28 +60,34 @@ public class ExcelSerializer<S extends Solution<S,I>, I extends Instance>  exten
     private final Optional<ExcelCustomizer> excelCustomizer;
     private final ExcelConfig config;
 
+    private final InstanceManager<I> instanceManager;
+
+    private Object[][] instancePropertyData;
+
     /**
      * Create an Excel serializer
      *
-     * @param serializerConfig excel serializer configuration
-     * @param solverConfig solver configuration
+     * @param serializerConfig         excel serializer configuration
+     * @param solverConfig             solver configuration
      * @param referenceResultProviders reference result providers if available
-     * @param excelCustomizer customizer if available
+     * @param excelCustomizer          customizer if available
+     * @param instanceManager
      */
     public ExcelSerializer(
             ExcelConfig serializerConfig,
             SolverConfig solverConfig,
             List<ReferenceResultProvider> referenceResultProviders,
-            Optional<ExcelCustomizer> excelCustomizer
-    ) {
+            Optional<ExcelCustomizer> excelCustomizer,
+            InstanceManager<I> instanceManager) {
         super(serializerConfig, referenceResultProviders);
         this.config = serializerConfig;
         this.maximizing = solverConfig.isMaximizing();
         this.excelCustomizer = excelCustomizer;
+        this.instanceManager = instanceManager;
     }
 
     /** {@inheritDoc} */
-    public void _serializeResults(List<SolutionGeneratedEvent<S, I>> results, Path p) {
+    public void _serializeResults(String experimentName, List<SolutionGeneratedEvent<S, I>> results, Path p) {
         log.info("Exporting result data to XLSX...");
 
         File f = p.toFile();
@@ -83,10 +98,14 @@ public class ExcelSerializer<S extends Solution<S,I>, I extends Instance>  exten
             var pivotSheet = excelBook.createSheet(PIVOT_SHEET);
             var rawSheet = excelBook.createSheet(RAW_SHEET);
 
-            RawSheetWriter writer = getRawSheetWriter(results);;
+            RawSheetWriter writer = getRawSheetWriter(results);
             var area = writer.fillRawSheet(rawSheet, maximizing, results, referenceResultProviders);
             
             fillPivotSheet(pivotSheet, area, rawSheet);
+
+            // Check and fill instance sheet if appropiate
+            fillInstanceSheet(experimentName, excelBook);
+
             if(this.excelCustomizer.isPresent()){
                 var realExcelCustomizer = excelCustomizer.get();
                 log.info("Calling Excel customizer: " + realExcelCustomizer.getClass().getSimpleName());
@@ -101,6 +120,54 @@ public class ExcelSerializer<S extends Solution<S,I>, I extends Instance>  exten
         } catch (Exception e) {
             throw new RuntimeException(String.format("Exception while trying to save Excel file: %s, reason: %s", f.getAbsolutePath(), e.getClass().getSimpleName()), e.getCause());
         }
+    }
+
+    protected void fillInstanceSheet(String expName, XSSFWorkbook excelBook) {
+        if(!this.config.isInstanceSheetEnabled()){
+            log.fine("Instance sheet disabled");
+            return;
+        }
+
+        log.fine("Creating instance sheet...");
+        var keys = Instance.getUniquePropertiesKeys().toArray(new String[0]);
+
+        if (keys.length == 0) {
+            log.info("Instance sheet enabled, but no data available, skipping");
+            this.instancePropertyData = new Object[][]{{"Instance sheet is enabled but no properties found for any instance, remember to define properties using Instance::setProperty. If you do not want to use this feature, set serializers.xlsx.instance-sheet-enabled to false"}};
+        } else {
+            if(this.instancePropertyData == null){
+                this.instancePropertyData = getInstancePropertyData(expName, keys);
+            }
+        }
+        var sheet = excelBook.createSheet(INSTANCE_SHEET);
+
+        // Write matrix data to cell Excel sheet
+        for (int i = 0; i < this.instancePropertyData.length; i++) {
+            var row = sheet.createRow(i);
+            var rowData = this.instancePropertyData[i];
+            for (int j = 0; j < rowData.length; j++) {
+                var cell = row.createCell(j);
+                writeCell(cell, rowData[j], RawSheetWriter.CType.VALUE);
+            }
+        }
+        log.fine("Instance sheet created");
+    }
+
+    private Object[][] getInstancePropertyData(String expName, String[] keys) {
+        List<Object[]> properties = new ArrayList<>();
+        Object[] header = new Object[keys.length + 1];
+        header[0] = "Instance name";
+        System.arraycopy(keys, 0, header, 1, keys.length);
+        properties.add(header);
+        instanceManager.getInstanceSolveOrder(expName).stream().map(instanceManager::getInstance).forEach(instance -> {
+            Object[] row = new Object[keys.length+1];
+            row[0] = instance.getId();
+            for (int i = 0; i < keys.length; i++) {
+                row[i+1] = instance.getPropertyOrDefault(keys[i], "");
+            }
+            properties.add(row);
+        });
+        return properties.toArray(new Object[0][]);
     }
 
     private RawSheetWriter getRawSheetWriter(List<? extends SolutionGeneratedEvent<S, I>> results) {